pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bpa-trading-system"
version = "0.1.0"
description = "Multi-exchange trading automation system"
authors = [
    {name = "Francisco Javier Lopez", email = "francisco.lopez4256@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "aiofiles==24.1.0",
    "aiohttp==3.12.15",
    "aiosqlite==0.21.0",
    "ccxt==4.5.5",
    "cepmex==1.0.0",
    "google-api-python-client==2.183.0",
    "google-auth-oauthlib==1.2.2",
    "ipinfo==5.2.1",
    "pandas==2.3.2",
    "pandas-ta==0.4.71b0",
    "pillow==11.3.0",
    "protobuf==6.32.1",
    "py-clob-client==0.25.0",
    "pytesseract==0.3.13",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/frank120121"
Repository = "https://github.com/frank120121/bpa-trading-system"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "exchanges*", "trading*", "data*", "utils*", "security*", "integrations*", "localization*"]
exclude = ["tests*", "docs*", "logs*", "config*"]